+++ ribs/emitters/_evolution_strategy_emitter.py
@@ -238,7 +238,7 @@ class EvolutionStrategyEmitter(EmitterBase):
                        self._batch_size // 2)
 
         # Update Evolution Strategy.
-        self._opt.tell(indices, num_parents)
+        self._opt.tell(indices, ranking_values, num_parents)
 
         # Check for reset.
         if (self._opt.check_stop(ranking_values[indices]) or